+++ app/src/components/pages/collectionsPage/collectionsPage.tsx
@@ -4,7 +4,6 @@ import {
   Box,
   Heading,
   HorizontalRule,
-  SearchBar,
   Menu,
   Pagination,
   Flex,
@@ -23,11 +22,13 @@ import {
 } from "@/src/config/constants";
 import { CollectionSearchManager } from "@/src/utils/searchManager";
 import { headerBreakpoints } from "@/src/utils/breakpoints";
+import DCSearchBar from "../../search/dcSearchBar";
 
 export function CollectionsPage({ data, collectionSearchParams }) {
   const { push } = useRouter();
   const pathname = usePathname();
   const headingRef = useRef<HTMLHeadingElement>(null);
+  const isFirstLoad = useRef<boolean>(false);
 
   const [isLoaded, setIsLoaded] = useState(false);
   const totalPages = totalNumPages(data.numResults, data.perPage);
@@ -52,8 +53,10 @@ export function CollectionsPage({ data, collectionSearchParams }) {
 
   useEffect(() => {
     setIsLoaded(true);
-    headingRef.current?.focus();
-
+    if (isFirstLoad.current) {
+      headingRef.current?.focus();
+    }
+    isFirstLoad.current = true;
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [collections]);
 
@@ -84,41 +87,12 @@ export function CollectionsPage({ data, collectionSearchParams }) {
           text="Collections"
           subtitle="Explore the New York Public Library's diverse collections, including digitized photographs, manuscripts, maps, and more. Start exploring by using the search bar below or browse through the collections."
         />
-        <SearchBar
-          sx={{
-            maxWidth: "462px",
-            flexFlow: "row nowrap",
-            button: {
-              borderRadius: "0px 2px 2px 0px",
-              "> svg": {
-                width: "14px",
-                height: "14px",
-              },
-              paddingTop: "xs",
-              paddingBottom: "xs",
-              paddingLeft: "s !important",
-              paddingRight: "s !important",
-              "> span": {
-                display: "block !important",
-              },
-            },
-            [`@media screen and (max-width: ${headerBreakpoints.lgMobile}px)`]:
-              {
-                button: {
-                  padding: "xs !important",
-                  gap: 0,
-                  "> span": {
-                    display: "none !important",
-                  },
-                  "> svg": {
-                    width: "18px",
-                    height: "18px",
-                  },
-                },
-              },
-          }}
-          id={"search-collections"}
+        <DCSearchBar
+          id="search-collections"
+          labelText="Search collections by title"
+          maxWrapperWidth="462px"
           textInputProps={{
+            id: "collections-search-text",
             isClearable: true,
             isClearableCallback: () =>
               collectionSearchManager.handleKeywordChange(DEFAULT_SEARCH_TERM),
@@ -134,8 +108,6 @@ export function CollectionsPage({ data, collectionSearchParams }) {
           onSubmit={() => {
             updateURL(collectionSearchManager.handleSearchSubmit());
           }}
-          labelText="Search collections by title"
-          aria-label="Search collections by title"
         />
       </Box>