+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added custom searchbar to handle search submit
 - Added new search context and `searchManager` class, implemented on `/collections` page (DR-3365)
 
 ## [0.3.1] 2025-01-30